### 2020-07-03 #2434 @whedon
[PRE REVIEW]: DetectorChecker: analyzing patterns of defects in detector screens
**Submitting author:** @tomaslaz (<a href="http://orcid.org/0000-0002-8351-9857">Tomas Lazauskas</a>)
**Repository:** <a href="https://github.com/alan-turing-institute/DetectorChecker" target ="_blank">https://github.com/alan-turing-institute/DetectorChecker</a>
**Version:** 1.0.8
