### 2020-07-09 #2434 @kyleniemeyer
Hi @tomaslaz, while we are finding an editor to handle your submission, could you revise your paper to include a Statement of Need? This is now a required section for JOSS articles, as described in our [documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Thanks!
