### 2020-07-09 #2434 @fboehm
Yes, I'm happy to be the editor for this submission.
