### 2020-07-12 #2474 @whedon
[REVIEW]: DetectorChecker: analyzing patterns of defects in detector screens
**Submitting author:** @tomaslaz (<a href="http://orcid.org/0000-0002-8351-9857">Tomas Lazauskas</a>)
**Repository:** <a href="https://github.com/alan-turing-institute/DetectorChecker" target ="_blank">https://github.com/alan-turing-institute/DetectorChecker</a>
**Version:** v1.1.0

@craddm , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
