### 2020-07-27 #2474 @tomaslaz
> @tomaslaz - it looks like you're missing a doi according to the examination from whedon.
>
> MISSING DOIs
>
> * https://doi.org/10.1038/146150e0 may be missing for title: Statistical method from the viewpoint of quality control
>
> Please be sure to verify that this doi is accurate, then add it to your manuscript.

The suggested DOI is for a paper, however, we cite a book.
