### 2020-07-27 #2474 @fboehm
Thanks for double-checking the doi, @tomaslaz. We'll proceed without a doi for that reference.
