### 2020-08-24 #2474 @fboehm
Thank you, @craddm for mentioning the issue. @tomaslaz - Do you have any questions on how to proceed with revisions based on the open issue from @craddm?

@janfreyberg Is there anything that I can do to help as you review the submission? Do you have any questions about the checklist? Thanks again!!
