### 2020-08-25 #2474 @janfreyberg
Hi all - apologies for the long delay in my review, I should have updated you earlier. I've just updated the checklist and created an issue on the DetectorChecker repo!
