### 2020-08-25 alan-turing-institute/DetectorChecker#111 @janfreyberg
The paper is great, the only thing missing in my view is a clearer summary of the field - what options do people currently have? If there is no other open-source package or no other package that implements this workflow from start to finish, it would be worth saying so.
