### 2020-11-09 d6364d3
Test that every file created by a plot exists
Add a convenience function to run the assertion that the file exists and remove the fie
