### 2020-11-09 ed42748
Test file exists in all tests. Move assertion util function to another file
