+++ tests/testthat/test_analysis.R
@@ -1,14 +1,5 @@
 context("Testing reading in the dead pixel data and visualizing the detector: Pilatus")
-
-#' Assert that a file exists and then remove it
-#'
-#' @param file_path A file path
-expect_file_exists <- function(file_path){
-
-  expect_true(file.exists(file_path))
-  if (file.exists(file_path)) file.remove(file_path)
-
-}
+source("utils.R")
 
 test_that("Pilatus", {
   test_dir <- getwd()
