+++ tests/testthat/test_pixel.R
@@ -101,10 +101,7 @@ test_that("Plotting pixel analysis", {
   plot_pixel_ctr_eucl(detector, file_path = test_out_path)
 
   # Check whether the file was created
-  expect_true(file.exists(test_out_path))
-
-  # Removing the test output file
-  if (file.exists(test_out_path)) file.remove(test_out_path)
+  expect_file_exists(test_out_path)
 
   # ----------------------------------------------------------------------------
   # Parallel maxima distances from the centre
@@ -115,10 +112,7 @@ test_that("Plotting pixel analysis", {
   plot_pixel_ctr_linf(detector, file_path = test_out_path)
 
   # Check whether the file was created
-  expect_true(file.exists(test_out_path))
-
-  # Removing the test output file
-  if (file.exists(test_out_path)) file.remove(test_out_path)
+  expect_file_exists(test_out_path)
 
   # ----------------------------------------------------------------------------
   # Pixel distances from corners
@@ -129,11 +123,7 @@ test_that("Plotting pixel analysis", {
   plot_pixel_dist_corner(detector, file_path = test_out_path)
 
   # Check whether the file was created
-  expect_true(file.exists(test_out_path))
-
-  # Removing the test output file
-  if (file.exists(test_out_path)) file.remove(test_out_path)
-
+  expect_file_exists(test_out_path)
   # ----------------------------------------------------------------------------
   # Pixel distances from the module edges by column
 
@@ -143,10 +133,7 @@ test_that("Plotting pixel analysis", {
   plot_pixel_dist_edge_col(detector, file_path = test_out_path)
 
   # Check whether the file was created
-  expect_true(file.exists(test_out_path))
-
-  # Removing the test output file
-  if (file.exists(test_out_path)) file.remove(test_out_path)
+  expect_file_exists(test_out_path)
 
   # ----------------------------------------------------------------------------
   # Pixel distances from the module edges by row
@@ -157,10 +144,7 @@ test_that("Plotting pixel analysis", {
   plot_pixel_dist_edge_row(detector, file_path = test_out_path)
 
   # Check whether the file was created
-  expect_true(file.exists(test_out_path))
-
-  # Removing the test output file
-  if (file.exists(test_out_path)) file.remove(test_out_path)
+  expect_file_exists(test_out_path)
 
   # ----------------------------------------------------------------------------
   # Pixel distances from the module edges
@@ -171,8 +155,5 @@ test_that("Plotting pixel analysis", {
   plot_pixel_dist_edge(detector, file_path = test_out_path)
 
   # Check whether the file was created
-  expect_true(file.exists(test_out_path))
-
-  # Removing the test output file
-  if (file.exists(test_out_path)) file.remove(test_out_path)
+  expect_file_exists(test_out_path)
 })
