### 2020-11-09 6c66c52
Make detector an S3 Class
