### 2020-11-09 92ba857
Replace plot_detector and summary_detector functions with S3 methods
Use S3 dispatch methods for summary and plot
