+++ DESCRIPTION
@@ -13,7 +13,7 @@ Description: DetectorChecker is an R package to aid in the assessment of damage 
 Depends: R (>= 3.6.2)
 Imports:
     spatstat (>= 1.63-3),
-    tools (>= 3.6.3), 
+    tools (>= 3.6.2), 
     tiff (>= 0.1-5), 
     hdf5r (>= 1.3.2), 
     raster (>= 3.1-5), 
