### 2020-11-09 525e4da
Add is.detector method
