### 2020-11-09 52db7a6
Tidy up docs
Dead_Stats is a list, not an S3 object
