+++ man/detector_summary.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/layout.R
-\name{detector_summary}
-\alias{detector_summary}
-\title{Generates a string with the detector summary}
-\usage{
-detector_summary(detector)
-}
-\arguments{
-\item{detector}{Detector object}
-}
-\value{
-String with the detector summary
-}
-\description{
-Generates a string with the detector summary
-}
