+++ man/plot.detector.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/layout.R
-\name{plot_detector}
-\alias{plot_detector}
+\name{plot.detector}
+\alias{plot.detector}
 \title{Plot detector}
 \usage{
-plot_detector(detector, file_path = NA, caption = TRUE)
+\method{plot}{detector}(detector, file_path = NA, caption = TRUE)
 }
 \arguments{
 \item{detector}{Detector object}
