+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(plot,detector)
+S3method(summary,detector)
 export(Dead_Stats)
 export(Excalibur_Detector)
 export(PerkinElmerCropped1600_Detector)
@@ -13,7 +15,6 @@ export(create_detector)
 export(dead_pix_coords)
 export(dead_stats_summary)
 export(detector_consist_check)
-export(detector_summary)
 export(dist_corner)
 export(dist_edge_col)
 export(dist_edge_min)
@@ -37,6 +38,7 @@ export(glm_pixel_dist_edge_col)
 export(glm_pixel_dist_edge_min)
 export(glm_pixel_dist_edge_row)
 export(ini_graphics)
+export(is.detector)
 export(load_pix_matrix)
 export(perform_glm)
 export(pixel_dist_ctr_eucl)
@@ -44,7 +46,6 @@ export(pixel_dist_ctr_linf)
 export(plot_angles)
 export(plot_arrows)
 export(plot_counts)
-export(plot_detector)
 export(plot_events)
 export(plot_events_angles)
 export(plot_events_arrows)
