### 2020-11-13 cec6d69
Make functions internal
