### 2020-11-13 7aca815
Make check_clumps internal
