### 2020-11-13 9bbb389
Make dead_pix_cords internal
