### 2020-11-13 f1eec0f
Update docs for detector creation
