### 2020-11-13 c588872
Make Dead_Stats internal
