### 2020-11-13 c150fa7
Actually Dead_Stats should be public to see data structure
