### 2020-11-13 bfb60c8
Update Dead_Stats docs
