### 2020-11-13 f9f0b82
Update dead_stats_summary docs and example
