### 2020-11-13 48630e6
Make detector_consist_check internal
