### 2020-11-13 1003e17
Update dist_corner docs
