### 2020-11-13 a61e8f4
Use package data in examples
