### 2020-11-13 f03c010
Make more functions internal
