+++ man/dot-get_clump_pixel_ppp.Rd
@@ -17,3 +17,4 @@ ppp object for damaged detector pixels
 \description{
 Creates ppp object for damaged detector pixels
 }
+\keyword{internal}
