+++ man/dot-ini_graphics.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/io.R
-\name{ini_graphics}
-\alias{ini_graphics}
+\name{.ini_graphics}
+\alias{.ini_graphics}
 \title{Starts the graphics device driver for producing graphics with respect to a
   chosen format}
 \usage{
-ini_graphics(file_path)
+.ini_graphics(file_path)
 }
 \arguments{
 \item{file_path}{Output path with an extension}
@@ -14,3 +14,4 @@ ini_graphics(file_path)
 Starts the graphics device driver for producing graphics with respect to a
   chosen format
 }
+\keyword{internal}
