+++ man/dot-create_ppp_edges_col.Rd
@@ -15,3 +15,4 @@ Point pattern dataset
 \description{
 This is the ppp_edges_col creation function
 }
+\keyword{internal}
