### 2020-11-17 8fbada8
Update more function docs
