+++ man/dot-plot_arrows.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plots.R
+\name{.plot_arrows}
+\alias{.plot_arrows}
+\title{Plots nearest neighbor oriented arrows}
+\usage{
+.plot_arrows(ppp_obj, caption, file_path = NA)
+}
+\arguments{
+\item{ppp_obj}{spatstat ppp object}
+
+\item{caption}{caption of the figure}
+
+\item{file_path}{file path}
+}
+\description{
+Plots nearest neighbor oriented arrows
+}
+\keyword{internal}
