+++ man/plot_angles.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/plots.R
-\name{plot_angles}
-\alias{plot_angles}
-\title{Plots NN angles}
-\usage{
-plot_angles(ppp_obj, caption, file_path = NA)
-}
-\arguments{
-\item{ppp_obj}{ppp object}
-
-\item{caption}{caption of the figure}
-
-\item{file_path}{file path}
-}
-\description{
-Plots NN angles
-}

[…]
