+++ R/analysis.R
@@ -1,7 +1,4 @@
-#' @title Analysis module
-
-
-#' A list to represent dead pixels statistics summary
+#' A list to represent dead pixels statistics summary.
 #' Added to a detector by \code{get_dead_stats}.
 #'
 #' @param dead_n Total number of damaged pixels:
@@ -254,9 +251,17 @@ plot_pixels_arrows <- function(detector, file_path = NA, row = NA, col = NA,
 #'
 #' @param detector Detector object
 #' @return Fitted model
-#' @examples
-#' glm_pixel_ctr_eucl(PerkinElmerFull_exp_1)
 #' @export
+#' @examples
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(
+#'  detector = detector_pilatus, file_path = file_path)
+#' # Fit logistic regression model
+#' glm_pixel_ctr_eucl(detector_pilatus)
 glm_pixel_ctr_eucl <- function(detector) {
   dist <- pixel_dist_ctr_eucl(detector)
   pix_matrix <- detector$pix_matrix
@@ -275,7 +280,15 @@ glm_pixel_ctr_eucl <- function(detector) {
 #' @return Fitted model
 #' @export
 #' @examples
-#' glm_pixel_ctr_linf(PerkinElmerFull_exp_1)
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(
+#'  detector = detector_pilatus, file_path = file_path)
+#' # Fit logistic regression model
+#' glm_pixel_ctr_linf(detector_pilatus)
 glm_pixel_ctr_linf <- function(detector) {
   dist <- pixel_dist_ctr_linf(detector)
   pix_matrix <- detector$pix_matrix
@@ -294,7 +307,15 @@ glm_pixel_ctr_linf <- function(detector) {
 #' @return Fitted model
 #' @export
 #' @examples
-#' glm_pixel_dist_edge_col(PerkinElmerFull_exp_1)
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(
+#'  detector = detector_pilatus, file_path = file_path)
+#' # Fit logistic regression model
+#' glm_pixel_dist_edge_col(detector_pilatus)
 glm_pixel_dist_edge_col <- function(detector) {
   dist <- dist_edge_col(detector)
   pix_matrix <- detector$pix_matrix
@@ -313,7 +334,15 @@ glm_pixel_dist_edge_col <- function(detector) {
 #' @return Fitted model
 #' @export
 #' @examples
-#' glm_pixel_dist_edge_row(PerkinElmerFull_exp_1)
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(
+#'  detector = detector_pilatus, file_path = file_path)
+#' # Fit logistic regression model
+#' glm_pixel_dist_edge_row(detector_pilatus)
[…]
