+++ R/pixel.R
@@ -27,12 +27,14 @@
 
 # Pixel analysis functions -----------------------------------------------------
 
-#' Calculate euclidean distance from the centre of a module for each pixel
+#' Calculate euclidean distance from the center of a module for each pixel
 #'
 #' @param detector Detector object
 #' @return Matrix of euclidean distances
 #' @examples
-#' pixel_dist_ctr_eucl(PerkinElmerFull_Detector())
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' pixel_dist_ctr_eucl(detector_pilatus)
 #' @export
 pixel_dist_ctr_eucl <- function(detector) {
   dist <- matrix(NA, nrow = detector$detector_height, ncol = detector$detector_width)
@@ -56,7 +58,10 @@ pixel_dist_ctr_eucl <- function(detector) {
 #' @param detector Detector object
 #' @return Matrix of parallel maxima
 #' @examples
-#' pixel_dist_ctr_linf(PerkinElmerFull_Detector())
+#' @examples
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' pixel_dist_ctr_linf(detector_pilatus)
 #' @export
 pixel_dist_ctr_linf <- function(detector) {
   dist <- matrix(NA, nrow = detector$detector_height, ncol = detector$detector_width)
@@ -230,6 +235,14 @@ dist_edge_min <- function(detector) {
 #' @param detector Detector object
 #' @param file_path Output file path
 #' @export
+#' @examples
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
+#' plot_pixel_ctr_eucl(detector_pilatus)
 plot_pixel_ctr_eucl <- function(detector, file_path = NA) {
   dist <- pixel_dist_ctr_eucl(detector)
 
@@ -244,6 +257,14 @@ plot_pixel_ctr_eucl <- function(detector, file_path = NA) {
 #' @param detector Detector object
 #' @param file_path Output file path
 #' @export
+#' @examples
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
+#' plot_pixel_ctr_linf(detector_pilatus)
 plot_pixel_ctr_linf <- function(detector, file_path = NA) {
   dist <- pixel_dist_ctr_linf(detector)
 
@@ -258,6 +279,14 @@ plot_pixel_ctr_linf <- function(detector, file_path = NA) {
 #' @param detector Detector object
 #' @param file_path Output file path
 #' @export
+#' @examples
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
+#' plot_pixel_dist_corner(detector_pilatus)
 plot_pixel_dist_corner <- function(detector, file_path = NA) {
   dist <- dist_corner(detector)
 
@@ -272,6 +301,14 @@ plot_pixel_dist_corner <- function(detector, file_path = NA) {
 #' @param detector Detector object
[…]
