+++ NAMESPACE
@@ -63,6 +63,7 @@ export(plot_pixel_dist_edge_row)
 export(plot_pixels)
 export(plot_pixels_angles)
 export(plot_pixels_arrows)
+export(plot_pixels_count)
 export(plot_pixels_density)
 export(plot_pixels_kfg)
 export(readin_detector)
