### 2020-11-18 262ca8b
Fix final example errors
