+++ R/analysis.R
@@ -116,7 +116,7 @@ plot_module_pixels <- function(detector, col, row, file_path = NA,
 #' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
 #'                          package ="detectorchecker")
 #' detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
-#' detector_pilatus_damage < get_dead_stats(detector_pilatus)
+#' detector_pilatus_damage <- get_dead_stats(detector_pilatus)
 #' plot_pixels_count(detector_pilatus_damage)
 plot_pixels_count <- function(detector, file_path = NA, row = NA, col = NA,
                                   caption = TRUE) {
