+++ R/clump.R
@@ -312,6 +312,11 @@ find_clumps <- function(detector, row = NA, col = NA) {
 
 #' Plots damaged detector events
 #'
+#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' The integer values declare which damage events to include:
+#'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
+#'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
+#' 
 #' @param detector Detector object
 #' @param col Module column number
 #' @param row Module row number
@@ -381,6 +386,11 @@ plot_events <- function(detector, col = NA, row = NA, file_path = NA, caption = 
 }
 
 #' Plots damaged detector module events
+#' 
+#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' The integer values declare which damage events to include:
+#'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
+#'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
 #'
 #' @param detector Detector object
 #' @param col Module column number
@@ -441,7 +451,12 @@ plot_module_events <- function(detector, col, row, file_path = NA, caption = TRU
 }
 
 #' Creates ppp object for damaged detector events
-#'
+#' 
+#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' The integer values declare which damage events to include:
+#'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
+#'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
+#' 
 #' @param detector Detector object
 #' @param incl_event_list a list of events to be included
 #' @param height Detector height
@@ -476,7 +491,12 @@ plot_module_events <- function(detector, col, row, file_path = NA, caption = TRU
 }
 
 #' Creates ppp object for damaged detector pixels
-#'
+#' 
+#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' The integer values declare which damage events to include:
+#'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
+#'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
+#' 
 #' @param detector Detector object
 #' @param incl_event_list a list of events to be included
 #' @return ppp object for damaged detector pixels
@@ -501,8 +521,12 @@ plot_module_events <- function(detector, col, row, file_path = NA, caption = TRU
 #' Generates events matrix for selected events
 #'
 #' Generates events mask (a matrix with pixels as 0 and events as 1) indicating if a pixel is in an event
-#' as calculated by \code{find_clumps(detc)}
-#'
+#' as calculated by \code{find_clumps(detc)} 
+#' Note that the parameter incl_event_list is a list of integers from [1,8]. For all events this would be list(1,2,3,4,5,6,7,8).
+#' The integer values declare which damage events to include:
+#'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
+#'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
+#' 
 #' @param detector Detector object
 #' @param incl_event_list a list of events to be included
 #' @return Events matrix
[…]
