+++ man/glm_events_dist_edge_row.Rd
@@ -15,5 +15,8 @@ glm_events_dist_edge_row(detector, incl_event_list = NA)
 Fitted model
 }
 \description{
-Fits events distances from the module edges by row using glm
+Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+The integer values declare which damage events to include:
+     1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
+     6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
 }
