+++ man/plot_events_kfg.Rd
@@ -30,7 +30,10 @@ plot_events_kfg(
 \item{incl_event_list}{a list of events to be included}
 }
 \description{
-Plots K, F, G functions of a detector or module
+Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+The integer values declare which damage events to include:
+     1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
+     6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
 }
 \examples{
 detector_perkinfull <- create_detector("PerkinElmerFull")
