### 2020-11-23 3caac97
Add details to find_clumps
Also rebuild paper pdf
