+++ R/clump.R
@@ -252,14 +252,42 @@
   return(dataFrame)
 }
 
-#' Locates and classifies clumps of damaged pixels in a detector.
-#' Adds an events matrix attribute to a \code{detector}
+#' Locates and classifies clumps (Events) of damaged pixels in a detector.
+#' Adds an events matrix attribute to a \code{detector}.
+#' @description
+#' Clumps are a collection of neighboring dead pixels.
+#' Pixels are classified as a clumps if they share an edge.
+#' A mathematical definition can be found \href{https://warwick.ac.uk/fac/sci/statistics/crism/research/17-02/17-02w.pdf}{here}.
+#' See details for classifications types.
+#' @details
+#' | Classification | Description |
+#' | ---- | ---- |
+#' | Singlton | Single connecting pixel. |
+#' | Doublets | Two connecting pixels. |
+#' | Triplets | Three connecting pixels. |
+#' | Upper horizontal line | Horizontal line in upper half of detector. |
+#' | Lower horizontal line | Horizontal line in lower half of detector. |
+#' | Left vertical line| Vertical line on left side of detector. |
+#' | Right vertical line | Vertical line on right side of detector. |
+#' | Larger cluster | Any larger grouping of damaged pixels. |
+#'
+#' Lines have a pixel width of 1 but my include up to 10% stray adjacent pixels.
+#'
 #'
 #' @param detector Detector object
 #' @param row Module row number
 #' @param col Module column number
 #' @return Detector with events matrix attribute
 #' @export
+#' @examples
+#' # Create a detector
+#' detector_pilatus <- create_detector("Pilatus")
+#' # Load a pixel matrix
+#' file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
+#'                          package ="detectorchecker")
+#' detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
+#'  #Find clumps
+#' detector_pilatus_w_clumps <- find_clumps(detector_pilatus)
 find_clumps <- function(detector, row = NA, col = NA) {
 
   pixel_mask <- get_dead_pix_mask(detector)
@@ -316,7 +344,7 @@ find_clumps <- function(detector, row = NA, col = NA) {
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
 #'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
-#' 
+#'
 #' @param detector Detector object
 #' @param col Module column number
 #' @param row Module row number
@@ -386,7 +414,7 @@ plot_events <- function(detector, col = NA, row = NA, file_path = NA, caption = 
 }
 
 #' Plots damaged detector module events
-#' 
+#'
 #' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
@@ -451,12 +479,12 @@ plot_module_events <- function(detector, col, row, file_path = NA, caption = TRU
 }
 
 #' Creates ppp object for damaged detector events
-#' 
+#'
 #' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
 #'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
-#' 
+#'
 #' @param detector Detector object
[…]
