+++ R/clump.R
@@ -340,7 +340,7 @@ find_clumps <- function(detector, row = NA, col = NA) {
 
 #' Plots damaged detector events
 #'
-#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' Note that the parameter incl_event_list is a list of integers from 1-8.  For all events this would be list(1,2,3,4,5,6,7,8).
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
 #'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
@@ -415,7 +415,7 @@ plot_events <- function(detector, col = NA, row = NA, file_path = NA, caption = 
 
 #' Plots damaged detector module events
 #'
-#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' Note that the parameter incl_event_list is a list of integers from 1-8.  For all events this would be list(1,2,3,4,5,6,7,8).
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
 #'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
@@ -480,7 +480,7 @@ plot_module_events <- function(detector, col, row, file_path = NA, caption = TRU
 
 #' Creates ppp object for damaged detector events
 #'
-#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' Note that the parameter incl_event_list is a list of integers from 1-8.  For all events this would be list(1,2,3,4,5,6,7,8).
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
 #'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
@@ -520,7 +520,7 @@ plot_module_events <- function(detector, col, row, file_path = NA, caption = TRU
 
 #' Creates ppp object for damaged detector pixels
 #'
-#' Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
+#' Note that the parameter incl_event_list is a list of integers from 1-8.  For all events this would be list(1,2,3,4,5,6,7,8).
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
 #'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
@@ -550,7 +550,7 @@ plot_module_events <- function(detector, col, row, file_path = NA, caption = TRU
 #'
 #' Generates events mask (a matrix with pixels as 0 and events as 1) indicating if a pixel is in an event
 #' as calculated by \code{find_clumps(detc)}
-#' Note that the parameter incl_event_list is a list of integers from [1,8]. For all events this would be list(1,2,3,4,5,6,7,8).
+#' Note that the parameter incl_event_list is a list of integers from 1-8. For all events this would be list(1,2,3,4,5,6,7,8).
 #' The integer values declare which damage events to include:
 #'      1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
 #'      6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
@@ -622,7 +622,7 @@ get_events_matrix <- function(detector, incl_event_list = NA) {
 
[…]
