### 2020-11-23 49925d0
Update clump classification
