### 2020-11-23 #2474 @OscartGiles
Dear @fboehm , sorry about the delay. I think the changes are now complete.

Thanks very much @fboehm and @craddm for your comments, they were really useful. Hope all is well Matt! (we used to be in the same department at UoL).

I have created a single PR with all the changes (https://github.com/alan-turing-institute/DetectorChecker/pull/116) but the description links to individual PRs which have more details about the changes and which issues they address.

How do progress from here @fboehm? Would you like the review the PRs before they are merged to master? There are small changes to the paper (@tomaslaz - you might need to resubmit). I'm also going to make a few minor changes to the WebApp now to correct for any changes in the R package's API.

All the best,

Oscar
