### 2020-12-05 a60f72a
Update paper.bib
added bib for shiny
