+++ JOSS-paper/paper.bib
@@ -116,6 +116,14 @@ URL = {https://improvement.nhs.uk/documents/6119/Transforming_imaging_services.p
   version      = {1.0.6},
 }
 
+@Manual{shiny,
+    title = {shiny: Web Application Framework for R},
+    author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
+    year = {2020},
+    note = {R package version 1.5.0},
+    url = {https://CRAN.R-project.org/package=shiny},
+  }
+
 @software{tomas_lazauskas_2020_3662235,
   author       = {T. Lazauskas and
                   O. T. Giles and
