### 2020-12-05 #2474 @fboehm
@OscartGiles - I've opened a pull request with a small addition; namely, I think that there should be a formal citation to Shiny. I've added the bib entry and placed the pandoc citation where it's needed. Once it's merged, I'll regenerate the pdf.
