### 2020-12-09 #2474 @tomaslaz
@fboehm thank you. The version number is v1.1.0 and the doi is 10.5281/zenodo.4312870

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4312870.svg)](https://doi.org/10.5281/zenodo.4312870)
