### 2020-12-09 #2474 @fboehm
The whedon output with missing DOIs is for a journal article and not for the book that the authors cite.
