### 2020-12-11 #2474 @Kevin-Mattheus-Moerman
@tomaslaz I am processing the final steps to accept this work in JOSS. There are some minor points that need work.

- [ ] Please amend the archive metadata such that the title and author list and order match that of the paper (@fboehm you can check this for future reference).

- [ ] Could you please add city information to your affiliations and also please spell out UK as United Kingdom?
