### 2020-12-17 #2474 @tomaslaz
> @tomaslaz I am processing the final steps to accept this work in JOSS. There are some minor points that need work.
>
> * [x]  Please amend the archive metadata such that the title and author list and order match that of the paper (@fboehm you can check this for future reference).
> * [x]  Could you please add city information to your affiliations and also please spell out UK as United Kingdom?

This has been done @Kevin-Mattheus-Moerman.
