### 2020-12-21 #2474 @Kevin-Mattheus-Moerman
@tomaslaz it looks like you have not completed point 2 of my request :point_up: in relation to the affiliations. It may feel double/excessive to add town names and spell out the country but we'd like to be more homogeneous in our approach across submissions.
